requirements.txt
fastapi
uvicorn
python-dotenv
scikit-learn
pdfminer.six
python-docx
groq
python-multipart
psycopg2-binary